add NameLogicTest
